# web_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DropTracker Web API",
    "version": "1.0.0",
    "description": "Versioned, browser-oriented JSON API consumed by the Next.js BFF. Separate from the RuneLite intake API. See FRONTEND_PLAN.md §6."
  },
  "servers": [{ "url": "/api/v1" }],
  "tags": [
    { "name": "leaderboards" },
    { "name": "players" },
    { "name": "groups" },
    { "name": "announcements" },
    { "name": "me" }
  ],
  "paths": {
    "/leaderboards/players": {
      "get": {
        "tags": ["leaderboards"],
        "operationId": "getPlayerLeaderboard",
        "summary": "Player loot leaderboard",
        "parameters": [
          { "$ref": "#/components/parameters/Period" },
          { "$ref": "#/components/parameters/Scope" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "A page of leaderboard entries.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LeaderboardPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/leaderboards/groups": {
      "get": {
        "tags": ["leaderboards"],
        "operationId": "getGroupLeaderboard",
        "summary": "Group loot leaderboard",
        "parameters": [
          { "$ref": "#/components/parameters/Period" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "A page of group leaderboard entries.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LeaderboardPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/players/{playerId}": {
      "get": {
        "tags": ["players"],
        "operationId": "getPlayer",
        "summary": "Player profile",
        "parameters": [{ "$ref": "#/components/parameters/PlayerId" }],
        "responses": {
          "200": {
            "description": "Player profile payload.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlayerProfile" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/players/search": {
      "get": {
        "tags": ["players"],
        "operationId": "searchPlayers",
        "summary": "Search players by name",
        "parameters": [{ "$ref": "#/components/parameters/Query" }],
        "responses": {
          "200": {
            "description": "Matching players.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PlayerSummary" }
                }
              }
            }
          }
        }
      }
    },
    "/groups/{groupId}": {
      "get": {
        "tags": ["groups"],
        "operationId": "getGroup",
        "summary": "Group profile",
        "parameters": [{ "$ref": "#/components/parameters/GroupId" }],
        "responses": {
          "200": {
            "description": "Group profile payload.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GroupProfile" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/announcements": {
      "get": {
        "tags": ["announcements"],
        "operationId": "listAnnouncements",
        "summary": "List announcements",
        "parameters": [
          { "$ref": "#/components/parameters/Scope" },
          { "$ref": "#/components/parameters/Cursor" }
        ],
        "responses": {
          "200": {
            "description": "A page of announcements.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementPage" }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["me"],
        "operationId": "getMe",
        "summary": "Current session: user, linked players, group roles",
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user payload.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Me" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dt_session"
      }
    },
    "parameters": {
      "Period": {
        "name": "period",
        "in": "query",
        "description": "Time partition: YYYYMM (monthly), YYYYWW (weekly), YYYYMMDD (daily) or 'all'.",
        "schema": { "type": "string", "default": "all" }
      },
      "Scope": {
        "name": "scope",
        "in": "query",
        "description": "global | group:{id} | npc:{id} | player:{id}",
        "schema": { "type": "string", "default": "global" }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": { "type": "string" }
      },
      "Query": {
        "name": "q",
        "in": "query",
        "required": true,
        "schema": { "type": "string", "minLength": 1 }
      },
      "PlayerId": {
        "name": "playerId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "GroupId": {
        "name": "groupId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "responses": {
      "Error": {
        "description": "RFC-7807 problem details.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC-7807 error envelope.",
        "required": ["title", "status"],
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      },
      "Money": {
        "type": "object",
        "description": "GP value: raw integer plus a preformatted display string.",
        "required": ["value", "value_formatted"],
        "properties": {
          "value": { "type": "integer", "format": "int64" },
          "value_formatted": { "type": "string", "examples": ["123.5M"] }
        }
      },
      "PageMeta": {
        "type": "object",
        "required": ["page", "limit", "total"],
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "LeaderboardEntry": {
        "type": "object",
        "required": ["rank", "id", "name", "loot"],
        "properties": {
          "rank": { "type": "integer" },
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "loot": { "$ref": "#/components/schemas/Money" },
          "delta": { "type": "integer", "description": "Recent change in GP (for live deltas)." }
        }
      },
      "LeaderboardPage": {
        "type": "object",
        "required": ["period", "scope", "entries", "meta"],
        "properties": {
          "period": { "type": "string" },
          "scope": { "type": "string" },
          "entries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LeaderboardEntry" }
          },
          "meta": { "$ref": "#/components/schemas/PageMeta" }
        }
      },
      "PlayerSummary": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "global_rank": { "type": "integer" },
          "total_loot": { "$ref": "#/components/schemas/Money" }
        }
      },
      "Submission": {
        "type": "object",
        "required": ["id", "type", "label", "ts"],
        "properties": {
          "id": { "type": "integer" },
          "type": {
            "type": "string",
            "enum": ["drop", "clog", "pb", "ca", "pet", "level", "quest"]
          },
          "label": { "type": "string" },
          "value": { "$ref": "#/components/schemas/Money" },
          "image_url": { "type": "string" },
          "ts": { "type": "integer", "description": "Unix seconds." }
        }
      },
      "GroupMembership": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" }
        }
      },
      "PlayerProfile": {
        "allOf": [
          { "$ref": "#/components/schemas/PlayerSummary" },
          {
            "type": "object",
            "properties": {
              "points": { "type": "integer" },
              "top_npc": { "type": "string" },
              "groups": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/GroupMembership" }
              },
              "recent_submissions": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Submission" }
              }
            }
          }
        ]
      },
      "GroupProfile": {
        "type": "object",
        "required": ["id", "name", "member_count"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "member_count": { "type": "integer" },
          "global_rank": { "type": "integer" },
          "monthly_loot": { "$ref": "#/components/schemas/Money" },
          "discord_url": { "type": "string" },
          "top_player": { "$ref": "#/components/schemas/PlayerSummary" },
          "recent_submissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Submission" }
          }
        }
      },
      "Announcement": {
        "type": "object",
        "required": ["id", "scope_type", "title", "body_md", "published_at"],
        "properties": {
          "id": { "type": "integer" },
          "scope_type": { "type": "string", "enum": ["global", "group"] },
          "group_id": { "type": "integer", "nullable": true },
          "title": { "type": "string" },
          "body_md": { "type": "string" },
          "cover_image_url": { "type": "string", "nullable": true },
          "pinned": { "type": "boolean" },
          "author_name": { "type": "string" },
          "published_at": { "type": "integer", "description": "Unix seconds." }
        }
      },
      "AnnouncementPage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Announcement" }
          },
          "next_cursor": { "type": "string", "nullable": true }
        }
      },
      "Me": {
        "type": "object",
        "required": ["user_id", "discord_id", "players", "groups"],
        "properties": {
          "user_id": { "type": "integer" },
          "discord_id": { "type": "string" },
          "display_name": { "type": "string" },
          "avatar_url": { "type": "string", "nullable": true },
          "players": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PlayerSummary" }
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "name", "role"],
              "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" },
                "role": {
                  "type": "string",
                  "enum": ["owner", "admin", "member"]
                }
              }
            }
          }
        }
      }
    }
  }
}